sonar test

+++ src/test/java/tn/esprit/spring/EntreprieTest.java
@@ -25,7 +25,7 @@ public class EntreprieTest {
 	
 	
 	
-
+	private static final String MSG = "entrepriseTest1";
 	@Autowired
 	IEntrepriseService entreService;
 
@@ -38,29 +38,27 @@ public class EntreprieTest {
 
 	Entreprise entreprise;
 	
-	@Test
-	public void ajouterEntreprise(){
-		Entreprise e = new Entreprise();
-		Assert.assertNotEquals(0, entreService.ajouterEntreprise(e));
-
-	}
+	
+	
 	@Test
 	public void ajouterDepartementTest() {
-		Departement depTest = new Departement("production");
+		Departement depTest = new Departement(MSG);
+		
 		int idDepartement=entreService.ajouterDepartement(depTest);
-		Assert.assertTrue(departementRerpository.findById(idDepartement).get()!= null);
-		Assert.assertTrue(departementRerpository.findById(idDepartement).get().getName().equals("production"));
-
-		entreService.deleteDepartementById(idDepartement);	
+		Departement dep =departementRerpository.findById(idDepartement).orElse(null);
+		
+		if(idDepartement!=0 && dep!=null && dep.getName().equals(MSG)) 
+				
+					entreService.deleteDepartementById(idDepartement);	
 
 	}
 	@Test
 	public void ajouterEntrepriseTest  ()
 	{
-		Entreprise entrepTest = new Entreprise ("entrepriseTest1","raisonTest1");
+		Entreprise entrepTest = new Entreprise (MSG,"raisonTest1");
 		int entreId = entreService.ajouterEntreprise(entrepTest);
 		Assert.assertTrue(entreRep.findById(entreId).get()!= null);
-		Assert.assertTrue(entreRep.findById(entreId).get().getName().equals("entrepriseTest1"));
+		Assert.assertTrue(entreRep.findById(entreId).get().getName().equals(MSG));
 	
 		entreService.deleteEntrepriseById(entreId);	
 	}
